refactor: rename BBButton to BBTooltipButton
