+++ frontend/src/bbkit/BBTooltipButton.vue
